Refactor build binaries stage. Extract duplicate code to the template
